requirements.txt
# Django Core
Django==5.2.7
django-shortuuidfield==0.1.3

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# API schema / docs
drf-spectacular==0.27.2

# Stripe + config + email backend
stripe==11.2.0
python-decouple==3.8
sendgrid==6.11.0
django-sendgrid-v5==1.2.4

# Filtering
django-filter==23.2

# Image Processing
Pillow==12.0.0

# HTTP Client (for testing)
requests==2.32.5

# JWT Token Handling
PyJWT==2.10.1

# Short UUID Generation
shortuuid==1.0.13

# Additional Dependencies
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
six==1.17.0
